Use XSD from https, rather than from http
This is a workaround for:
https://www.w3.org/blog/2022/07/redirecting-to-https-on-www-w3-org/:

25 July 2022

W3C’s main web site www.w3.org has been available via https for over a decade,
but until now we have not been redirecting all requests to https
as is commonly done on most other sites.

The primary reason for this is that we wanted to avoid causing issues
for software requesting machine-readable resources from www.w3.org
such as HTML DTDs, XML Schemas, and namespace documents.

We believe enough time has passed for most such software to have been updated
to handle redirects and https, so we are planning to start redirecting
all requests received over http to https within a month or two.

In order to discover any potential remaining issues and to give
some advance notice in case there are software systems that still have issues
with redirects and https, we plan to conduct some limited tests
before fully deploying this change to our site,
where we redirect all http traffic to https for a few hours at a time.

The first such test is planned for Monday August 1,
for 8 hours starting at 14:00 UTC (14:00 UTC to 22:00 UTC)